Use Hugging Face Space for navatar generation
